Refactoring: Spielzeit mittels Game-Ticks von physikalischer Zeit abstrahiert - Die Spielzeit wird nun vom Game als 0-based Millisekunden pro Spiel verwaltet. Context hat keine SDL-Uhrzeit mehr. - Gamelogik aus Main-Loop in Game::update() verlagert. - Game::speed kann nun nur noch ganzzahlige Werte annehmen

+++ src/cpp/economics/EconomicsMgr.h
@@ -31,9 +31,10 @@ struct FindBuildingToGetGoodsFromResult {
     GoodsSlot goodsSlot;
 
     /**
-     * @brief SDL_GetTicks-Wert, wann zuletzt eine Abholung der Waren bei diesem Gebäude stattgefunden hat
+     * @brief [Spielzeitpunkt](@ref gameTicks), wann zuletzt eine Abholung der Waren bei diesem Gebäude
+     * stattgefunden hat
      */
-    unsigned int lastGoodsCollections;
+    unsigned long lastGoodsCollections;
 
     /**
      * leerer Konstruktur, der ein "nix zu holen"-Ergebnis anlegt
